Added an error catching handler to mvn_kp_read that will take care of most errors that occur in it, or procedures/functions it calls.
Added how errors shoudl be handled in mvn_kp_filename_parsers and mvn_kp_iuvs_filename.pro which is:
"on_error, 1"
This will print the stack and return to the calling function, instead of the
default which is to halt.

ALso added a debug option to all of these procedures. THe debug option will be
how what users can provide if they would like the code to halt in place on
an error instead of returning to main.
This debug option gets passed down from the top (mvn_kp_read in this case) to
all of the user defined procedures.

We will need to all procedures that we want errors handled in this manner to take the debug option as well as pass it on to procedures they call.