1. Added final CDF read/write routines. Added CDF master files, which are necessary for creating CDF files.
2. Updated mvn_kp_read.pro to read cdf files by default, /savefiles is now the option to use save files, /textfiles is the option
to read ascii format. There is no longer a "/binary" flag.

3. Pulled out some "configuration" information, like filename conventions, and put those into mvn_kp_config.pro. That procedure
is then called from the various routines that need that info. In this way, if a filename changes, or anythign else we pull out into
the config file, it can be changed in once place.

4. Added a new routine "mvn_kp_compare_data.pro" which compares two structures (or arrays of structrures) of either insitu or
iuvs like data structures. It will compare them element by element and error out at any point it finds something different. THis
was helpful for verifying my conversion to CDF was creating the same datat we had in save files. It will also be helpful for adding
in more regression tests. I also moved "regression_test" to testing_tools/ directory.

5. I'm adding "mvn_kp_create_insitu_save.pro" which will take in a list of insitu text files, and output save files.

6. Finally, I stripped out some of the IUVS read logic into a procedure called "mvn_kp_read_iuvs_file.pro".