Moved setup/teardown preference logic into testsuite_uts__define
Also added logic for checking output of mvn_kp_insitu_search,insitu, /list
by adding a journal file to capture stdout. I then added a known "good" version
of this to be compared to.

I added the journal file name as an environment variable in the init method
so any unit test can use it. It then gets removed in the cleanup method
if still around.